Setting up session list

+++ app/src/main/java/com/kjipo/bluetoothmidi/session/MidiSessionRepository.kt
@@ -4,7 +4,7 @@ import com.kjipo.bluetoothmidi.midi.MidiMessage
 
 interface MidiSessionRepository {
 
-    suspend fun startSession(): Int
+    suspend fun startSession(): Long
 
     suspend fun addMessageToSession(translatedMidiMessage: MidiMessage)
